backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.27
pydantic==2.6.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
mysql-connector-python==8.2.0 